+++ src/cards/fulcro/democards/card_ui.cljs
@@ -1,10 +1,12 @@
 (ns fulcro.democards.card-ui
   (:require
     [clojure.spec.test.alpha :as st]
+    fulcro.democards.alpha.dom-cards
     fulcro.democards.load-cards
+    fulcro.democards.i18n-alpha-cards
     fulcro.democards.manual-tests-of-dynamic-queries
     fulcro.democards.react-refs
     fulcro.democards.root-form-refresh
     fulcro.democards.parent-state-refresh))
 
-;(st/instrument)
+(st/instrument)